% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{add_alpha_to_color}
\alias{add_alpha_to_color}
\title{Add Alpha Level to a Passed Color}
\usage{
add_alpha_to_color(color, alpha)
}
\arguments{
\item{color}{Expected to be a color produced by rgb or compatible}

\item{alpha}{Expected to be a value between 0-1}
}
\description{
This function takes a color produced by 'rgb' in hex format and adds an alpha value.
}
\examples{
red = rgb(1,0,0);
red_alpha = add_alpha_to_color(red,0.5);
}
\keyword{alpha}

