% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{add_under_color}
\alias{add_under_color}
\title{Add A Color Underneath a Set of X-Y Coordinates}
\usage{
add_under_color(x, y, color)
}
\arguments{
\item{x}{X-coordindates of the line}

\item{y}{X-coordindates of the line}

\item{color}{Color of the shading underneath the line}
}
\description{
This function adds shading underneath a line, extending down to the x-axis
}
\examples{
plot(0:5,0:5,type='l',col=rgb(1,0,0))
add_under_color(0:5,0:5,rgb(1,0,0,0.5));
}
\keyword{line}
\keyword{plot}

