% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{applyBetterParSettings}
\alias{applyBetterParSettings}
\title{Apply a better set of ploting settings}
\usage{
applyBetterParSettings()
}
\description{
This function draws a bar plot with confidence intervals from a provided list
containing your data.
}

