% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{boxplot.fancy}
\alias{boxplot.fancy}
\title{Make a boxplot with some small improvements}
\usage{
\method{boxplot}{fancy}(data, add.N.count = T, label.names = NA, ...)
}
\arguments{
\item{...}{Optional: Any additional parameters will be passed on to the plotLinesWithConfInt function}

\item{data:}{A list containing your data sets}

\item{add.N.count:}{Whether or not to add the N counts to the data labels, defaults to True}
}
\description{
Make a boxplot with some small improvements
}
\examples{
boxplot.fancy(list(A=rnorm(10),B=rnorm(10)))
}
\keyword{boxplot}
\keyword{fancy}

