% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{buildTimelapsePlotWithConfInt}
\alias{buildTimelapsePlotWithConfInt}
\title{Make a Time-lapse ggplot2 Object with Confidence Intervals}
\usage{
buildTimelapsePlotWithConfInt(inputData, x.tick.position = NA,
  time.interval = 1, conf.int = 0.95, xlabel = NA, ylabel = NA,
  legend.labels = NA)
}
\arguments{
\item{inputData}{A list of matrices containing your data sets}

\item{x.tick.position}{Optional: The location of the x axis numeric
positions, if not specified, defaults to time.interval}

\item{time.interval}{Optional: The amount of time between each sample,
defaults to 1}

\item{conf.int}{Optional: Percentage size of the confidence interval plot,
defaults to 95%}

\item{xlabel}{Optional: If specified, will be passed to xlab}

\item{ylabel}{Optional: If specified, will be passed to ylab}

\item{legend.labels}{Optional: A mapping between the names of the inputData 
list and the desired legend titles}
}
\description{
This function builds a ggplot2 object that plots the mean of a provided set 
of data and the corresponding confidence interval that data at each time 
point. The input format is expected to be a list of matrices where each 
matrix represents a set of time series. Each of the matrices in the list are 
formated so that each row is a single time point and each column is a single 
experiment.
}
\details{
The confidence interval is calculated using a t.test and indicated on the 
plot as a shaded area that matches the color of the mean line.
}
\examples{
data = list(A = matrix(1:10,nrow=10,ncol=10) + rnorm(100), 
            B = matrix(1:10,nrow=10,ncol=10) + rnorm(100) + 5,
            C = matrix(1:10,nrow=10,ncol=10) + rnorm(100,sd=5));
thisPlot = buildTimelapsePlotWithConfInt(data);
thisPlot
}
\keyword{confidence}
\keyword{interval}
\keyword{plot}
