% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{convertSVGtoPNG}
\alias{convertSVGtoPNG}
\title{Convert a SVG image to PNG using imagemagick}
\usage{
convertSVGtoPNG(svg.file.name, im.width = 1000, debug = F)
}
\arguments{
\item{im.width}{Optional: Specifies the output width of the png file, defaults
to 1000}

\item{debug}{Optional: Print out imagemagick convert command}

\item{svg.file.name:}{The location of the svg file}
}
\description{
This function takes in an svg file name and converts that image to a png. 
This function assumes that imagemagick has been installed and that "convert"
is available at the command line. This function also assumes that the svg file
has the string '.svg' present at the end of the of the svg file
}
\examples{
svg.file = 'test.svg';
svg(svg.file)
plot(1:10)
graphics.off()
convertSVGtoPNG(svg.file)
}
\keyword{PNG}
\keyword{SVG}
\keyword{convert}

