% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{convertSVGtoTarget}
\alias{convertSVGtoTarget}
\title{Convert a SVG image to different image type using imagemagick}
\usage{
convertSVGtoTarget(svg.file.name, im.width = 1000, target = "jpg",
  debug = F)
}
\arguments{
\item{im.width}{Optional: Specifies the output width of the output file,
defaults to 1000}

\item{target}{Optional: The extension of the output file, defaults to .jpg}

\item{debug}{Optional: Print out imagemagick convert command}

\item{svg.file.name:}{The location of the svg file}
}
\description{
This function takes in an svg file name and converts that image to a
different format using imagemagick. This function assumes that imagemagick
has been installed and that "convert" is available at the command line. This
function also assumes that the svg file has the string '.svg' present at the
end of the of the svg file
}
\examples{
svg.file = 'test.svg';
svg(svg.file)
plot(1:10)
graphics.off()
convertSVGtoTarget(svg.file,target="jpg")
}
\keyword{SVG}
\keyword{convert}
