% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{lines_with_undershading}
\alias{lines_with_undershading}
\title{Draws A Line with Shading Underneath}
\usage{
lines_with_undershading(x, y, color, undercolor = NA, ...)
}
\arguments{
\item{x}{X-coordindates of the line}

\item{y}{X-coordindates of the line}

\item{color}{Color of the shading underneath the line}

\item{undercolor}{Optional: if specified, the shading underneath the line will be this color, if unspecified it defaults to the same color as the line, with alpha level of 0.25}

\item{...}{Optional: Any additional parameters will be passed on to the lines function}
}
\description{
This function adds shading underneath a line, extending down to the x-axis
}
\examples{
lines_with_undershading(0:5,0:5,col=rgb(1,0,0))
}
\keyword{line}
\keyword{plot}

