% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{plotBarplotWithConfInt}
\alias{plotBarplotWithConfInt}
\title{Draw a Barplot with Confidence Intervals}
\usage{
plotBarplotWithConfInt(data, label_names = NA, padj = NA, conf.int = 0.95,
  ...)
}
\arguments{
\item{...}{Optional: Any additional parameters will be passed on to the plotLinesWithConfInt function}

\item{mat}{Matrix of data}

\item{x}{Optional: X-coordindates of the line}

\item{color}{Optional: Change the color of the plotted points/CI shading}
}
\description{
This function draws a line with confidence interval shading, assuming all data in one column should be used to build the confidence interval
}
\examples{
plotLinesMatWithConfInt(...)
}
\keyword{confidence}
\keyword{interval}
\keyword{plot}

