% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{plotLinesMatWithConfInt}
\alias{plotLinesMatWithConfInt}
\title{Draws A Line with Confidence Interval Shading From a Collection of Data}
\usage{
plotLinesMatWithConfInt(mat, x_coords = NA, color = NA, ...)
}
\arguments{
\item{mat}{Matrix of data}

\item{color}{Optional: Change the color of the plotted points/CI shading}

\item{...}{Optional: Any additional parameters will be passed on to the plotLinesWithConfInt function}

\item{x}{Optional: X-coordindates of the line}
}
\description{
This function draws a line with confidence interval shading, assuming all data in one column should be used to build the confidence interval
}
\examples{
plotLinesMatWithConfInt(...)
}
\keyword{confidence}
\keyword{interval}
\keyword{plot}

