% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{plotLinesWithConfInt}
\alias{plotLinesWithConfInt}
\title{Draws A Line with Confidence Interval Shading}
\usage{
plotLinesWithConfInt(x, y, top_interval, bottom_interval, color = NA, ...)
}
\arguments{
\item{x}{X-coordindates of the line}

\item{y}{X-coordindates of the line}

\item{top_interval}{Y-coordindates of the top of the interval}

\item{bottom_interval}{Y-coordindates of the bottom of the interval}

\item{color}{Optional: Change the color of the plotted points/CI shading}

\item{...}{Optional: Any additional parameters will be passed on to the lines function}
}
\description{
This function draws a line with confidence interval shading
}
\examples{
plot_with_conf_int(1:5,1:5,c(1:5)+1,c(1:5)-1,rgb(0.5,0.5,0.5,0.5),typ='l')
}
\keyword{confidence}
\keyword{interval}
\keyword{plot}

