% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_misc.R
\name{read.csv.all}
\alias{read.csv.all}
\title{Read all CSV files found in search string}
\usage{
read.csv.all(search_str)
}
\arguments{
\item{search_str}{string to pass to Sys.glob for the search}
}
\description{
This function uses Sys.glob to search for files matching a given string and then cycles through those files, reading them in as CSVs. This function assumes that all your CSV files are the same format or at least can be 'rbind'ed together.
}
\examples{
all_CSVs = read.csv.all('directory/*/*.csv')
}
\keyword{CSV}
