% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{theme_berginski}
\alias{theme_berginski}
\title{Define a ggplot2 theme for myself}
\usage{
theme_berginski()
}
\description{
This function returns a theme that can be directly used in ggplot2 plots.
}
