% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{trimImage}
\alias{trimImage}
\title{Trim the whitespace around an image using Imagemagick}
\usage{
trimImage(file.name, debug = F)
}
\arguments{
\item{debug}{Optional: Print out imagemagick convert command}

\item{file.name:}{The location of the file}
}
\description{
This function takes a filename for an image, which will then be processed using the trim command in imagemagick to remove the surrounding whitespace
}
\examples{
\dontrun{
file.name = 'test.png';
svg(file.name)
plot(1:10)
graphics.off()
trimImage(file.name)
}
}
\keyword{trim}
