% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{applyBetterParSettings}
\alias{applyBetterParSettings}
\title{Apply a better set of ploting settings}
\usage{
applyBetterParSettings(...)
}
\description{
This function sets sets the par settings to something I think is a bit more 
reasonable for base R graphics. Any of these settings can be overriden by
passing in different values
}
\details{
The settings and explanation are:

-bty='n': remove the box around the entire plot

-mgp=c(1.5,0.5,0): reduce distance of the axis title from the axis (1.5), 
reduce distance of tick mark labels from the ticks (0.5)

-mar=c(3,3,0,0): reduce the margins around the (bottom,left,top,right) of the
plot
}
