% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_misc.R
\name{plotBarplotWithConfInt}
\alias{plotBarplotWithConfInt}
\title{Draw a Barplot with Confidence Intervals}
\usage{
plotBarplotWithConfInt(data, label_names = NA, padj = NA, conf.int = 0.95,
  add.N.count = FALSE, ...)
}
\arguments{
\item{data}{A list containing your data sets}

\item{label_names}{Optional: Labels for your data sets, if not provided the 
names of the data parameter will be used}

\item{padj}{Optional: Shift the position of the bar labels, negative numbers 
go up while positive numbers go down, defaults to NA}

\item{conf.int}{Optional: The confidence interval percentile, defaults to 
0.95}

\item{add.N.count}{Optional: Add the number of observations to the bar label
names}

\item{...}{Optional: Any additional parameters will be passed on to the 
barplot function}
}
\description{
This function draws a bar plot with confidence intervals from a provided list
containing your data.
}
\examples{
data = list(A = seq(1,5),B=seq(1,10));
plotBarplotWithConfInt(data)
}
\keyword{confidence}
\keyword{interval}
\keyword{plot}

